Fixed Bugs
Added a status variable that keeps track of whether or not a projectile has collided with the edge of the screen or enemy. Beforehand, the projectiles would flash between two different states (due to synch issues) and they would stay on screen after initially colliding with an enemy or screen edge. This status variable (and associated updates between players) fixes that issue.

Now the only issue is connecting between two different machines. The game is able to work if run on the same machine, but two different machines are unable to connect. Further information will be gathered to fix this issue.