Update v0.03
Added Player upgrades to the end of each wave. Further ones will be added but the current list is as follows:

        ("Increased Movement Speed", "+10% Base Movement Speed")
        ("Enhanced Fire Rate", "Decreases the delay between shots by 10%")
        ("Extended Health", "+10% Base Health Points")
        ("Armor Upgrade", "Reduces damage taken by 10%")
        ("Extended Stamina", "+10% Maximum Stamina")
        ("Health Regeneration", "Slowly regenerates health over time")
        ("Stamina Regeneration Boost", "Increases stamina regeneration rate by 10%")

Additionally, a stats tracker has been added to the left side of the screen. There are a few small bugs that will need to be fixed (such as picking up a fire rate powerup resets the fire rate upgrades).